requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
bokeh==3.3.4
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.8
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
Flask==3.1.0
fonttools==4.44.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.1
numpy==1.26.1
openai==1.58.1
packaging==23.2
pandas==2.2.0
Pillow==10.1.0
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
Werkzeug==3.1.3
wordcloud==1.9.4
xyzservices==2023.10.1
